refactored front end to allow multiple views for different pages within the app. added bare bones profile page
